+++ authpass/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixes #188 - Show banner to users to suggest activating autofill (@ATofighi)
 * KDBX: Fix merging of files with incoming deleted objects.
 * Add support for permanently deleting groups and password entries. #99
+* When a conflict is detected during save, automatically merge files #80
 
 # 1.8.2 - 2021-07-22